# Conflicts:
#	app/src/main/java/eu/faircode/email/AdapterMessage.java

+++ app/src/main/java/eu/faircode/email/FragmentMessages.java
@@ -107,7 +107,6 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.snackbar.Snackbar;
 import com.sun.mail.util.FolderClosedIOException;
 
-import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.openintents.openpgp.OpenPgpError;
@@ -3870,8 +3869,7 @@ public class FragmentMessages extends FragmentBase implements SharedPreferences.
                                 db.beginTransaction();
 
                                 // Write decrypted body
-                                Helper.writeText(message.getFile(context),
-                                        decrypted.toString().replace("\0", ""));
+                                Helper.writeText(message.getFile(context), decrypted.toString());
 
                                 db.message().setMessageStored(id, new Date().getTime());
 
@@ -3894,8 +3892,6 @@ public class FragmentMessages extends FragmentBase implements SharedPreferences.
 
                                 // Write decrypted body
                                 String html = parts.getHtml(context);
-                                if (html != null)
-                                    html = html.replace("\0", "");
                                 Helper.writeText(message.getFile(context), html);
 
                                 // Remove previously decrypted attachments
@@ -3992,6 +3988,10 @@ public class FragmentMessages extends FragmentBase implements SharedPreferences.
                     if (message == null)
                         return null;
 
+                    EntityAccount account = db.account().getAccount(message.account);
+                    if (account == null)
+                        return null;
+
                     EntityFolder folder = db.folder().getFolder(message.folder);
                     if (folder == null)
                         return null;
@@ -4006,7 +4006,7 @@ public class FragmentMessages extends FragmentBase implements SharedPreferences.
                             NotificationManager nm = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
                             nm.cancel("send:" + message.identity, 1);
                         }
-                    } else if (message.uid == null) {
+                    } else if (message.uid == null && !account.pop) {
                         db.message().deleteMessage(id);
                         db.folder().setFolderError(message.folder, null);
                     } else
@@ -4332,7 +4332,7 @@ public class FragmentMessages extends FragmentBase implements SharedPreferences.
                     return null;
 
                 String html = Helper.readText(file);
-                Document document = Jsoup.parse(html);
+                Document document = JsoupEx.parse(html);
                 HtmlHelper.embedImages(context, id, document);
 
                 Element body = document.body();
